+++ .env
@@ -4,4 +4,5 @@ NEXT_PUBLIC_FIREBASE_PROJECT_ID = truckers-95774
 NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET = truckers-95774.appspot.com
 NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID = 1073817008230
 NEXT_PUBLIC_FIREBASE_API_KEY = AIzaSyBF3_iJro9OnasNJI3Z46QmZCH1lmUTt8M
-REACT_APP_GOOGLE_MAPS_API_KEY = AIzaSyCxF5mBRVIcJmmnpnvfzlejI2N79H7PZf4
+# REACT_APP_GOOGLE_MAPS_API_KEY = AIzaSyCxF5mBRVIcJmmnpnvfzlejI2N79H7PZf4
+NEXT_PUBLIC_GOOGLE_MAPS_API_KEY = AIzaSyDtgjIl85v3STNcvO9InTMy_4XfDp_VoL8
