+++ pages/Reports.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { FaArrowCircleRight } from "react-icons/fa";
-import { AiOutlineArrowRight } from "react-icons/ai";
+
 import { GrDocumentUpload } from "react-icons/gr";
 import { getFirestore, collection, getDocs } from "firebase/firestore";
-// import PostLoader from "../Components/Post/Loader.tsx";
+import ReportCard from "../Components/ReportCard";
 import {
   Flex,
   Spacer,
@@ -19,13 +19,6 @@ import {
   VStack,
   Wrap,
   WrapItem,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-  ModalFooter,
   useDisclosure,
   Spinner,
 } from "@chakra-ui/react";
@@ -38,138 +31,9 @@ import ModalWrapper from "../Components/Modal/ModalWrapper.tsx";
 
 import Footer from "../Components/Footer";
 
-function CardModal({ imageSrc, heading, description, time, location }) {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  return (
-    <>
-      <Button
-        onClick={onOpen}
-        variant={"link"}
-        colorScheme={"blue"}
-        size={"sm"}
-      >
-        Read More
-        <AiOutlineArrowRight style={{ marginLeft: "5px" }} />
-      </Button>
-      <Modal
-        isCentered
-        onClose={onClose}
-        isOpen={isOpen}
-        motionPreset="slideInBottom"
-      >
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>{heading}</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <VStack
-            // maxW={{ base: "full", md: "275px" }}
-            // w={"lg"}
-            // borderWidth="1px"
-            // borderRadius="10px"
-            // overflow="hidden"
-            // // pt={"64px"}
-            // m="32px"
-            // // mx="auto">
-            >
-              <Stack align="start" spacing="2">
-                <HStack
-                  fontSize="16px"
-                  fontWeight="400"
-                  color="#5F6D7E"
-                  lineHeight="10"
-                >
-                  <Text>{location}:</Text>
-                  <Text>{time}:</Text>
-                  {/* <Text>{timeStamp.date}</Text> */}
-                </HStack>
-                <Flex align={"center"} justify={"center"} objectFit="contain">
-                  <Image
-                    src={imageSrc}
-                    alt={"Image description"}
-                    width={"lg"}
-                  />
-                </Flex>
-                <Box>
-                  <Heading
-                    fontSize="32"
-                    lineHeight="30px"
-                    fontWeight="600"
-                    size="md"
-                    style={{ marginTop: "35px" }}
-                  ></Heading>
-
-                  <br />
-                </Box>
-              </Stack>
-            </VStack>
-          </ModalBody>
-          <ModalFooter>
-            <Button colorScheme="blue" mr={3} onClick={onClose}>
-              Close
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
-    </>
-  );
-}
-
-const Card = ({ heading, time, description, location, imageSrc, href }) => {
-  return (
-    <Box
-      w={{ md: "lg" }}
-      borderWidth="1px"
-      borderRadius="10px"
-      overflow="hidden"
-      m={{ base: "4", md: "auto" }}
-    >
-      <Stack align="start" spacing="2">
-        <Flex align={"center"} justify={"center"} objectFit="contain">
-          <Image src={imageSrc} alt={"Image description"} width={"lg"} />
-        </Flex>
-        <Box>
-          <Heading
-            fontSize="32"
-            lineHeight="30px"
-            fontWeight="600"
-            size="md"
-            style={{ marginTop: "35px" }}
-          >
-            {/* {heading} */}
-            <Text>{heading}</Text>
-          </Heading>
-
-          <HStack
-            fontSize="16px"
-            fontWeight="400"
-            color="#5F6D7E"
-            lineHeight="10"
-          >
-            <Text>{location}:</Text>
-            <Text>{time}:</Text>
-            {/* <Text>{timeStamp.date}</Text> */}
-          </HStack>
-
-          <br />
-        </Box>
-
-        <CardModal
-          imageSrc={imageSrc}
-          heading={heading}
-          description={description}
-          time={time}
-          location={location}
-        ></CardModal>
-      </Stack>
-    </Box>
-    // </>
-  );
-};
-
 // create a loading indicator
 
-const LoadingWidget = () => {
+export const LoadingWidget = () => {
   return (
     <Box
       height="100vh"
@@ -201,6 +65,9 @@ export default function Report() {
       });
 
       setReports(allReports);
+      if (allReports.length == 0) {
+        setIsLoading(false);
+      }
       console.log("All reports:", allReports, "isLoading:", isLoading);
     } catch (error) {
       console.error("Error retrieving reports:", error);
@@ -214,7 +81,6 @@ export default function Report() {
     console.log("All reports:", reports.length);
     if (reports.length > 0) {
       setIsLoading(!isLoading);
-      console.log(reports.length, isLoading);
     }
   }, [reports]);
 
@@ -256,7 +122,7 @@ export default function Report() {
         {reports.map((report, idx) => {
           return (
             <WrapItem key={idx}>
-              <Card
+              <ReportCard
                 heading={report.caption}
                 imageSrc={report.attachment}
                 date={report.date}
